Fix callback creation and usage

+++ src/config/config.rs
@@ -12,7 +12,7 @@ pub struct Config {
     pub saved_directory: Option<PathBuf>,
     pub playlists: Vec<PlaylistInfo>,
     pub codec: CodecPreference,
-    pub max_concurrent_downloads: usize,
+    pub max_concurrent_downloads: i32,
 }
 
 #[derive(Serialize, Debug, Deserialize, Clone, Default)]
@@ -83,6 +83,7 @@ impl Config {
         Ok(playlist_id)
     }
 }
+
 
 async fn save_config(
     config: &Config,
